+++ CHANGES.md
@@ -1,6 +1,7 @@
 ## Version 3.0.4 - 2024/07/??
 
 ### Added
+- The Architect Table in the Garrison will now automatically designate a plot for the Barracks, provided that the player's Garrison is at level 1 and they are currently on the "Build Your Barracks" quest.
 ### Changed or Updated
 ### Deprecated
 ### Removed